[REFACTOR] spk init command
