+++ src/commands/deployment/dashboard.test.ts
@@ -3,12 +3,12 @@ jest.mock("../../config");
 
 import { Config } from "../../config";
 import { exec } from "../../lib/shell";
+import { validatePrereqs } from "../../lib/validator";
 import {
   disableVerboseLogging,
   enableVerboseLogging,
   logger
 } from "../../logger";
-import { validatePrereqs } from "../init";
 import {
   extractManifestRepositoryInformation,
   getEnvVars,
