+++ src/config.test.ts
@@ -2,7 +2,16 @@ import os from "os";
 import path from "path";
 import shell from "shelljs";
 import uuid from "uuid/v4";
-import { Bedrock, updateVariableWithLocalEnv, write } from "./config";
+import {
+  Bedrock,
+  Config,
+  defaultConfigFile,
+  loadConfiguration,
+  saveConfiguration,
+  updateVariableWithLocalEnv,
+  write
+} from "./config";
+import { createTempDir } from "./lib/ioUtil";
 import { disableVerboseLogging, enableVerboseLogging } from "./logger";
 import { IBedrockFile } from "./types";
 
@@ -126,5 +135,76 @@ describe("Bedrock", () => {
     }
     expect(bedrockConfig).toBeFalsy();
     expect(error).toBeDefined();
+  });
+});
+
+const mockFileName = "src/commands/mocks/spk-config.yaml";
+
+describe("Initializing a project to use spk with a config file", () => {
+  test("init command basic file test", async () => {
+    // Create random directory to initialize
+    const randomTmpDir = createTempDir();
+    process.env.test_name = "my_storage_account";
+    process.env.test_key = "my_storage_key";
+    const filename = path.resolve(mockFileName);
+    await saveConfiguration(filename, randomTmpDir);
+    loadConfiguration(path.join(randomTmpDir, "config.yaml"));
+
+    const config = Config();
+    expect(config.introspection!).toBeDefined();
+    expect(config.introspection!.azure!.account_name).toBe(
+      process.env.test_name
+    );
+    const key = await config.introspection!.azure!.key;
+    expect(key).toBe(process.env.test_key);
+    expect(config.introspection!.azure!.table_name!).toBe(
+      process.env.test_name + "+" + process.env.test_key
+    );
+  });
+});
+
+describe("Initializing a project a config file but no env vars", () => {
+  test("init command basic file without env vars", async () => {
+    const filename = path.resolve(mockFileName);
+    process.env.test_name = "";
+    process.env.test_key = "";
+    try {
+      loadConfiguration(filename);
+      // Make sure execution does not get here:
+      expect(true).toBeFalsy();
+    } catch (err) {
+      expect(err).toBeDefined();
+    }
+  });
+});
+
+describe("Initializing a project with a non-existent file", () => {
+  test("Non-existent file test", async () => {
+    const filename = path.resolve("./spk-config-test.yaml");
+    try {
+      loadConfiguration(filename);
+      // Make sure execution does not get here:
+      expect(true).toBeFalsy();
+    } catch (e) {
+      expect(e.code).toBe("ENOENT");
+    }
+  });
+});
+
+describe("Writing to default config location", () => {
+  test("Default config location exists", async () => {
+    try {
+      const filename = path.resolve(mockFileName);
+      process.env.test_name = "testStorageName";
+      process.env.test_key = "testStorageKey";
+      loadConfiguration(filename);
+
+      await saveConfiguration(filename);
+      loadConfiguration(defaultConfigFile());
+      expect(Config().azure_devops!).toBeDefined();
+    } catch (e) {
+      // Make sure execution does not get here:
+      expect(true).toBeFalsy();
+    }
   });
 });
