+++ src/index.ts
@@ -18,7 +18,7 @@ import { deploymentCommand } from "./commands/deployment";
 import { hldCommand } from "./commands/hld";
 import { infraCommand } from "./commands/infra";
 import { ingressCommand } from "./commands/ingress-route";
-import { initCommandDecorator } from "./commands/init";
+import { commandDecorator as initCommandDecorator } from "./commands/init";
 import { projectCommand } from "./commands/project";
 import { serviceCommand } from "./commands/service";
 import { variableGroupCommand } from "./commands/variable-group";
